Migrate from Quasar CLI webpack to Vite
- Replace @quasar/app-webpack with @quasar/app-vite v2.0.0
- Upgrade Quasar to v2.16.11 and Vue to v3.4.38
- Add Pinia v2.3.1 and pinia-plugin-persistedstate v3.2.3
- Upgrade @obiba/quasar-app-extension-amber to v1.2.0
- Remove webpack-specific dependencies (eslint-webpack-plugin, workbox-webpack-plugin, core-js)
- Add Vite build tools (postcss, postcss-rtlcss, sass-embedded)
- Remove babel.config.js (not needed with Vite)
- Convert quasar.conf.js to quasar.config.js with ES modules
- Replace chainWebpack with extendViteConf
- Create postcss.config.cjs for RTL support
- Create index.html with Vite format
- Update FontAwesome from v5 to v7 (required by @quasar/extras v2.0.0)
- Fix Vue component imports to include .vue extensions
- Remove browserslist from package.json (Vite uses different browser targeting)

Build successfully completes with Vite
