--- .postcssrc.js
@@ -1,8 +1,0 @@
-// https://github.com/michael-ciniawsky/postcss-load-config
-
-module.exports = {
-  plugins: [
-    // to edit target browsers: use "browserslist" field in package.json
-    require('autoprefixer')
-  ]
-}
