--- babel.config.js
@@ -1,15 +1,0 @@
-
-
-module.exports = api => {
-  return {
-    presets: [
-      [
-        '@quasar/babel-preset-app',
-        api.caller(caller => caller && caller.target === 'node')
-          ? { targets: { node: 'current' } }
-          : {}
-      ]
-    ]
-  }
-}
-
