+++ package.json
@@ -19,22 +19,21 @@
     "@feathersjs/feathers": "^4.5.11",
     "@feathersjs/rest-client": "^4.5.11",
     "@feathersjs/vuex": "^4.0.1-pre.7",
-    "@quasar/app-webpack": "^3.12.3",
-    "@quasar/cli": "^2.3.0",
-    "@quasar/extras": "^1.16.9",
+    "@quasar/extras": "^2.0.0",
     "@vuelidate/core": "^2.0.0-alpha.32",
     "@vuelidate/validators": "^2.0.0-alpha.25",
     "axios": "^0.21.1",
     "bcryptjs": "^2.4.3",
-    "core-js": "^3.6.5",
     "crypto": "^1.0.1",
     "feathers-hooks-common": "^5.0.6",
     "marked": "^9.1.1",
     "merge-anything": "^5.0.2",
-    "quasar": "^2.14.3",
+    "pinia": "^2.3.1",
+    "pinia-plugin-persistedstate": "^3.2.3",
+    "quasar": "^2.16.11",
     "secure-ls": "^1.2.6",
     "uuid": "^8.3.2",
-    "vue": "^3.0.0",
+    "vue": "^3.4.38",
     "vue-i18n": "^9.0.0",
     "vue-recaptcha-v3": "^2.0.1",
     "vue-router": "^4.0.0",
@@ -43,28 +42,19 @@
   },
   "devDependencies": {
     "@babel/eslint-parser": "^7.13.14",
-    "@obiba/quasar-app-extension-amber": "1.1.6",
+    "@obiba/quasar-app-extension-amber": "1.2.0",
+    "@quasar/app-vite": "^2.0.0",
     "eslint": "^7.14.0",
     "eslint-config-standard": "^16.0.2",
     "eslint-plugin-import": "^2.19.1",
     "eslint-plugin-node": "^11.0.0",
     "eslint-plugin-promise": "^5.1.0",
     "eslint-plugin-vue": "^7.0.0",
-    "eslint-webpack-plugin": "^2.4.0",
-    "workbox-webpack-plugin": "^6.3.0",
+    "postcss": "^8.5.8",
+    "postcss-rtlcss": "^5.4.0",
+    "sass-embedded": "^1.93.2",
     "yarn-audit-fix": "^10.0.7"
   },
-  "browserslist": [
-    "last 10 Chrome versions",
-    "last 10 Firefox versions",
-    "last 4 Edge versions",
-    "last 7 Safari versions",
-    "last 8 Android versions",
-    "last 8 ChromeAndroid versions",
-    "last 8 FirefoxAndroid versions",
-    "last 10 iOS versions",
-    "last 5 Opera versions"
-  ],
   "engines": {
     "node": ">= 12.22.1",
     "npm": ">= 6.13.4"
