+++ postcss.config.cjs
@@ -1,0 +1,9 @@
+// https://github.com/michael-ciniawsky/postcss-load-config
+
+const rtlcss = require('postcss-rtlcss')
+
+module.exports = {
+  plugins: [
+    rtlcss()
+  ]
+}
